Initial commit with State merge to Context class (#11)
