+++ CMakeLists.txt
@@ -115,7 +115,6 @@ add_executable(Dynamite
     ${CMAKE_SOURCE_DIR}/example/Dynamite/editor.cpp
     ${CMAKE_SOURCE_DIR}/example/Dynamite/palette.cpp
     ${CMAKE_SOURCE_DIR}/example/Dynamite/multipanel.cpp
-    ${CMAKE_SOURCE_DIR}/example/Dynamite/state.cpp
     ${CMAKE_SOURCE_DIR}/example/Dynamite/dyndsp_wrapper.cpp
     )
 target_link_libraries(Dynamite imnodes SDL2::SDL2)
